requirements.txt
django==1.11.*
psycopg2-binary==2.8.*
python-decouple==3.4
dj-config-url~=0.1.1